moved methods from concent to apiconsent
