+++ .github/workflows/build_artifacts.yaml
@@ -1,6 +1,6 @@
 name: build
 on:
-#  push:
+  push:
 
   workflow_dispatch:
